Time: 59 ms (38.95%), Space: 63.6 MB (5.07%) - LeetHub

+++ 2537-count-the-number-of-good-subarrays/2537-count-the-number-of-good-subarrays.java
@@ -1,16 +1,16 @@
 class Solution {
     public long countGood(int[] nums, int k) {
         Map<Integer,Integer> map = new HashMap<>();
-        int cnt=0, res=0;
+        long cnt=0, res=0;
         int n=nums.length,l=0;
         for(int r=0;r<n;r++){
             map.put(nums[r], map.getOrDefault(nums[r],0)+1);
-            if(map.get(nums[r])>1) cnt+=map.get(nums[r])-1;
+            if(map.get(nums[r])>1) cnt+=(long)map.get(nums[r])-1;
             while(cnt>=k){
-                if(map.get(nums[l])>1) cnt-=map.get(nums[l])-1;
+                if(map.get(nums[l])>1) cnt-=(long)map.get(nums[l])-1;
                 map.put(nums[l],map.get(nums[l])-1);
                 l++;
-                res+=n-r;
+                res+=(long)(n-r);
             }
         }
         return res;
